+++ crates/p2p/src/commands.rs
@@ -33,7 +33,7 @@ pub enum Command {
         /// Key for DHT record: a remote node's application public key.
         app_public_key: PublicKey,
         /// One shot channel for sending result back.
-        tx_back: tokio::sync::oneshot::Sender<Option<RecordData>>,
+        response_sender: tokio::sync::oneshot::Sender<Option<RecordData>>,
     },
 }
