+++ crates/p2p/src/swarm/setup/upgrade.rs
@@ -14,7 +14,7 @@ use libp2p::{
 
 use crate::{
     signer::ApplicationSigner,
-    swarm::{errors::SetupUpgradeError, message::SignedMessage},
+    swarm::{errors::SetupUpgradeError, signed_data::SignedData},
 };
 
 /// Inbound upgrade for handling incoming setup requests.
@@ -40,13 +40,13 @@ impl UpgradeInfo for InboundSetupUpgrade {
 }
 
 impl InboundUpgrade<Stream> for InboundSetupUpgrade {
-    type Output = SignedMessage;
+    type Output = SignedData;
     type Error = SetupUpgradeError;
     type Future = Pin<Box<dyn Future<Output = Result<Self::Output, Self::Error>> + Send>>;
 
     fn upgrade_inbound(self, stream: Stream, _: Self::Info) -> Self::Future {
         Box::pin(async move {
-            let mut framed = Framed::new(stream, JsonCodec::<SignedMessage, SignedMessage>::new());
+            let mut framed = Framed::new(stream, JsonCodec::<SignedData, SignedData>::new());
 
             match StreamExt::next(&mut framed).await {
                 Some(Ok(signed_message)) => Ok(signed_message),
@@ -100,7 +100,7 @@ impl<S: ApplicationSigner> OutboundUpgrade<Stream> for OutboundSetupUpgrade<S> {
 
     fn upgrade_outbound(self, stream: Stream, _: Self::Info) -> Self::Future {
         Box::pin(async move {
-            let setup_message = SignedMessage::new_signed_setup(
+            let setup_message = SignedData::new_signed_setup(
                 self.app_public_key.clone(),
                 self.local_transport_id,
                 self.remote_transport_id,
@@ -108,7 +108,7 @@ impl<S: ApplicationSigner> OutboundUpgrade<Stream> for OutboundSetupUpgrade<S> {
             )
             .map_err(|e| SetupUpgradeError::SignedMessageCreation(e.into()))?;
 
-            let mut framed = Framed::new(stream, JsonCodec::<SignedMessage, SignedMessage>::new());
+            let mut framed = Framed::new(stream, JsonCodec::<SignedData, SignedData>::new());
             framed
                 .send(setup_message)
                 .await
